[FIX] service, model 계층 분리
